Added server side wrappers for S3Proxy based and filesystem based implementation. Moved the ACL management code to the server side.
